Drop support for PostgreSQL <9.3
