+++ .travis.yml
@@ -98,14 +98,6 @@ jobs:
     - stage: Test
       dist: trusty
       php: 7.3
-      env: DB=pgsql POSTGRESQL_VERSION=9.2 COVERAGE=yes
-      services:
-        - postgresql
-      addons:
-        postgresql: "9.2"
-    - stage: Test
-      dist: trusty
-      php: 7.3
       env: DB=pgsql POSTGRESQL_VERSION=9.3 COVERAGE=yes
       services:
         - postgresql
