+++ docs/en/reference/types.rst
@@ -781,9 +781,7 @@ Please also notice the mapping specific footnotes for additional information.
 |                   |               |                          |         +----------------------------------------------------------+
 |                   |               |                          |         | ``LONGTEXT`` [20]_                                       |
 |                   |               +--------------------------+---------+----------------------------------------------------------+
-|                   |               | **PostgreSQL**           | < 9.2   | ``TEXT`` [1]_                                            |
-|                   |               |                          +---------+----------------------------------------------------------+
-|                   |               |                          | < 9.4   | ``JSON``                                                 |
+|                   |               | **PostgreSQL**           | < 9.4   | ``JSON``                                                 |
 |                   |               |                          +---------+----------------------------------------------------------+
 |                   |               |                          | >= 9.4  | ``JSON`` [21]_                                           |
 |                   |               |                          |         +----------------------------------------------------------+
