+++ lib/Doctrine/DBAL/Driver/AbstractPostgreSQLDriver.php
@@ -6,8 +6,6 @@ use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
 use Doctrine\DBAL\Exception;
 use Doctrine\DBAL\Platforms\PostgreSQL100Platform;
-use Doctrine\DBAL\Platforms\PostgreSQL91Platform;
-use Doctrine\DBAL\Platforms\PostgreSQL92Platform;
 use Doctrine\DBAL\Platforms\PostgreSQL94Platform;
 use Doctrine\DBAL\Platforms\PostgreSqlPlatform;
 use Doctrine\DBAL\Schema\PostgreSqlSchemaManager;
@@ -100,10 +98,6 @@ abstract class AbstractPostgreSQLDriver implements ExceptionConverterDriver, Ver
                 return new PostgreSQL100Platform();
             case version_compare($version, '9.4', '>='):
                 return new PostgreSQL94Platform();
-            case version_compare($version, '9.2', '>='):
-                return new PostgreSQL92Platform();
-            case version_compare($version, '9.1', '>='):
-                return new PostgreSQL91Platform();
             default:
                 return new PostgreSqlPlatform();
         }
