--- lib/Doctrine/DBAL/Platforms/Keywords/PostgreSQL91Keywords.php
@@ -1,126 +1,0 @@
-<?php
-
-namespace Doctrine\DBAL\Platforms\Keywords;
-
-/**
- * PostgreSQL 9.1 reserved keywords list.
- */
-class PostgreSQL91Keywords extends PostgreSQLKeywords
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function getName()
-    {
-        return 'PostgreSQL91';
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @link http://www.postgresql.org/docs/9.1/static/sql-keywords-appendix.html
-     */
-    protected function getKeywords()
-    {
-        return [
-            'ALL',
-            'ANALYSE',
-            'ANALYZE',
-            'AND',
-            'ANY',
-            'ARRAY',
-            'AS',
-            'ASC',
-            'ASYMMETRIC',
-            'AUTHORIZATION',
-            'BINARY',
-            'BOTH',
-            'CASE',
-            'CAST',
-            'CHECK',
-            'COLLATE',
-            'COLUMN',
-            'CONCURRENTLY',
-            'CONSTRAINT',
-            'CREATE',
-            'CROSS',
-            'CURRENT_CATALOG',
-            'CURRENT_DATE',
-            'CURRENT_ROLE',
-            'CURRENT_SCHEMA',
-            'CURRENT_TIME',
-            'CURRENT_TIMESTAMP',
-            'CURRENT_USER',
-            'DEFAULT',
-            'DEFERRABLE',
-            'DESC',
-            'DISTINCT',
-            'DO',
-            'ELSE',
-            'END',
-            'EXCEPT',
-            'FALSE',
-            'FETCH',
-            'FOR',
-            'FOREIGN',
-            'FREEZE',
-            'FROM',
-            'FULL',
-            'GRANT',
-            'GROUP',
-            'HAVING',
-            'ILIKE',
-            'IN',
-            'INITIALLY',
-            'INNER',
-            'INTERSECT',
-            'INTO',
-            'IS',
-            'ISNULL',
-            'JOIN',
-            'LEADING',
-            'LEFT',
-            'LIKE',
-            'LIMIT',
-            'LOCALTIME',
-            'LOCALTIMESTAMP',
-            'NATURAL',
-            'NOT',
-            'NOTNULL',
-            'NULL',
-            'OFFSET',
-            'ON',
-            'ONLY',
-            'OR',
-            'ORDER',
-            'OUTER',
-            'OVER',
-            'OVERLAPS',
-            'PLACING',
-            'PRIMARY',
-            'REFERENCES',
-            'RETURNING',
-            'RIGHT',
-            'SELECT',
-            'SESSION_USER',
-            'SIMILAR',
-            'SOME',
-            'SYMMETRIC',
-            'TABLE',
-            'THEN',
-            'TO',
-            'TRAILING',
-            'TRUE',
-            'UNION',
-            'UNIQUE',
-            'USER',
-            'USING',
-            'VARIADIC',
-            'VERBOSE',
-            'WHEN',
-            'WHERE',
-            'WINDOW',
-            'WITH',
-        ];
-    }
-}
