+++ lib/Doctrine/DBAL/Platforms/Keywords/PostgreSQLKeywords.php
@@ -3,7 +3,7 @@
 namespace Doctrine\DBAL\Platforms\Keywords;
 
 /**
- * PostgreSQL Keywordlist.
+ * PostgreSQL 9.3 keywords list.
  */
 class PostgreSQLKeywords extends KeywordList
 {
@@ -26,20 +26,27 @@ class PostgreSQLKeywords extends KeywordList
             'ANALYZE',
             'AND',
             'ANY',
+            'ARRAY',
             'AS',
             'ASC',
+            'ASYMMETRIC',
             'AUTHORIZATION',
-            'BETWEEN',
             'BINARY',
             'BOTH',
             'CASE',
             'CAST',
             'CHECK',
             'COLLATE',
+            'COLLATION',
             'COLUMN',
+            'CONCURRENTLY',
             'CONSTRAINT',
             'CREATE',
+            'CROSS',
+            'CURRENT_CATALOG',
             'CURRENT_DATE',
+            'CURRENT_ROLE',
+            'CURRENT_SCHEMA',
             'CURRENT_TIME',
             'CURRENT_TIMESTAMP',
             'CURRENT_USER',
@@ -52,6 +59,7 @@ class PostgreSQLKeywords extends KeywordList
             'END',
             'EXCEPT',
             'FALSE',
+            'FETCH',
             'FOR',
             'FOREIGN',
             'FREEZE',
@@ -76,26 +84,27 @@ class PostgreSQLKeywords extends KeywordList
             'LOCALTIME',
             'LOCALTIMESTAMP',
             'NATURAL',
-            'NEW',
             'NOT',
             'NOTNULL',
             'NULL',
-            'OFF',
             'OFFSET',
-            'OLD',
             'ON',
             'ONLY',
             'OR',
             'ORDER',
             'OUTER',
+            'OVER',
             'OVERLAPS',
             'PLACING',
             'PRIMARY',
             'REFERENCES',
+            'RETURNING',
+            'RIGHT',
             'SELECT',
             'SESSION_USER',
             'SIMILAR',
             'SOME',
+            'SYMMETRIC',
             'TABLE',
             'THEN',
             'TO',
@@ -105,9 +114,12 @@ class PostgreSQLKeywords extends KeywordList
             'UNIQUE',
             'USER',
             'USING',
+            'VARIADIC',
             'VERBOSE',
             'WHEN',
             'WHERE',
+            'WINDOW',
+            'WITH',
         ];
     }
 }
