+++ lib/Doctrine/DBAL/Platforms/PostgreSQL94Platform.php
@@ -7,7 +7,7 @@ use Doctrine\DBAL\Types\Types;
 /**
  * Provides the behavior, features and SQL dialect of the PostgreSQL 9.4 database platform.
  */
-class PostgreSQL94Platform extends PostgreSQL92Platform
+class PostgreSQL94Platform extends PostgreSqlPlatform
 {
     /**
      * {@inheritdoc}
