+++ lib/Doctrine/DBAL/Platforms/PostgreSqlPlatform.php
@@ -33,9 +33,7 @@ use function strtolower;
 use function trim;
 
 /**
- * PostgreSqlPlatform.
- *
- * @deprecated Use PostgreSQL 9.4 or newer
+ * Provides the behavior, features and SQL dialect of the PostgreSQL 9.4+ database platform.
  *
  * @todo   Rename: PostgreSQLPlatform
  */
@@ -394,6 +392,7 @@ SQL
                     quote_ident(a.attname) AS field,
                     t.typname AS type,
                     format_type(a.atttypid, a.atttypmod) AS complete_type,
+                    (SELECT tc.collcollate FROM pg_catalog.pg_collation tc WHERE tc.oid = a.attcollation) AS collation,
                     (SELECT t1.typname FROM pg_catalog.pg_type t1 WHERE t1.oid = t.typbasetype) AS domain_type,
                     (SELECT format_type(t2.typbasetype, t2.typtypmod) FROM
                       pg_catalog.pg_type t2 WHERE t2.typtype = 'd' AND t2.oid = a.atttypid) AS domain_complete_type,
@@ -454,7 +453,7 @@ SQL
      */
     public function getCloseActiveDatabaseConnectionsSQL($database)
     {
-        return 'SELECT pg_terminate_backend(procpid) FROM pg_stat_activity WHERE datname = '
+        return 'SELECT pg_terminate_backend(pid) FROM pg_stat_activity WHERE datname = '
             . $this->quoteStringLiteral($database);
     }
 
@@ -975,6 +974,10 @@ SQL
      */
     public function getSmallIntTypeDeclarationSQL(array $field)
     {
+        if (! empty($field['autoincrement'])) {
+            return 'SMALLSERIAL';
+        }
+
         return 'SMALLINT';
     }
 
@@ -1163,6 +1166,7 @@ SQL
             'year'          => 'date',
             'uuid'          => 'guid',
             'bytea'         => 'blob',
+            'json'          => Type::JSON,
         ];
     }
 
@@ -1191,6 +1195,14 @@ SQL
     }
 
     /**
+     * {@inheritdoc}
+     */
+    public function hasNativeJsonType()
+    {
+        return true;
+    }
+
+    /**
      * {@inheritDoc}
      */
     protected function getReservedKeywordsClass()
@@ -1216,6 +1228,30 @@ SQL
         }
 
         return parent::getDefaultValueDeclarationSQL($field);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function supportsColumnCollation()
+    {
+        return true;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getColumnCollationDeclarationSQL($collation)
+    {
+        return 'COLLATE ' . $this->quoteSingleIdentifier($collation);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getJsonTypeDeclarationSQL(array $field)
+    {
+        return 'JSON';
     }
 
     /**
